include: gpio: Increase maximum GPIO number's value
For the Linux userspace platform, the GPIO number is composed by
the gpiochip base address + the actual GPIO number. The resulted
value can't be stored to an "uint8_t" in most of the cases.

+++ include/gpio.h
@@ -63,7 +63,7 @@
  */
 typedef struct gpio_init_param {
 	/** GPIO number */
-	uint8_t		number;
+	uint32_t	number;
 	/** GPIO extra parameters (device specific) */
 	void		*extra;
 } gpio_init_param;
@@ -74,7 +74,7 @@ typedef struct gpio_init_param {
  */
 typedef struct gpio_desc {
 	/** GPIO number */
-	uint8_t		number;
+	uint32_t	number;
 	/** GPIO extra parameters (device specific) */
 	void		*extra;
 } gpio_desc;
